% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_class}
\alias{comp_class}
\title{ICD10 code regroups for complication analysis}
\format{
An object of class \code{data.frame} with 84 rows and 3 columns.
}
\usage{
comp_class
}
\description{
ICD10 code regroups for complication analysis
}
\keyword{datasets}
