% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab_mapping}
\alias{lab_mapping}
\title{lab_mapping with nam and loinc}
\format{
An object of class \code{data.frame} with 20 rows and 6 columns.
}
\usage{
lab_mapping
}
\description{
lab_mapping with nam and loinc
}
\keyword{datasets}
