% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_ICD}
\alias{pheno_ICD}
\title{link between ICD and pheno}
\format{
An object of class \code{data.frame} with 25063 rows and 6 columns.
}
\usage{
pheno_ICD
}
\description{
link between ICD and pheno
}
\keyword{datasets}
