% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Runs the analytic workflow for the ards project}
\usage{
runAnalysis(obfusquation = TRUE, obfuscationThreeshord = 3)
}
\arguments{
\item{obfuscationThreeshord}{}
}
\value{

}
\description{
Runs the analytic workflow for the ards project
}
\keyword{4CE}
